Remove separate namespaces command and include it into the debug command
