+++ composer.json
@@ -18,6 +18,7 @@
     ],
     "require": {
         "php": "^7.2",
+        "ext-json": "*",
         "symfony/cache": "^4.2",
         "symfony/config": "^4.2",
         "symfony/console": "^4.2",
